pyproject.toml
[project]
name = "gnoma"
version = "0.1.0"
description = "Autonomous discovery of genetic and epigenetic interventions for reverse aging using reinforcement learning and biologically grounded world models"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = "MIT"

dependencies = [
    "cellxgene-census>=1.15.0",
    "scanpy>=1.10.0",
    "scvi-tools>=1.2.0",
    "anndata>=0.10.0",
    "gymnasium>=1.0.0",
    "stable-baselines3>=2.3.0",
    "torch>=2.2.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.17.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scipy>=1.12.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "umap-learn>=0.5.0",
    "tiledbsoma>=1.9.0",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "ipykernel>=6.29.0",
    "jupyterlab>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gnoma"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
